Homework #8 part 1 correct
